# [WIP] Unroll the loop, do more work during pipeline stall