This patch improves simdjson performance on AVX-2 by 3-4% in my tests. [This comment](https://github.com/lemire/simdjson/pull/310#issuecomment-530602731) describes the 4 changes made, and the reasons why. It also adds a -f option to parse.cpp so that causes it to only run find_structural_bits--this was critical to isolating the different performance gains.