@lemire 2019-09-09
Let us look at this!

There are 32 registers...