@jkeiser 2019-09-09
My reading of the Intel avx doc says it has 16 256-bit (YMM) registers, am I missing something there? Or is it just that most modern machines have enough of 512 implemented to have 32 registers now?