@lemire 2019-09-09
I mean 16 general purpose registers + 16 YMM registers... gives you 32.