@lemire 2019-09-09
I confirm that there is a surprisingly large positive effect to this PR, so we will undoubtably want to merge it.