@jkeiser 2019-09-09
Gotcha. The reason register pressure comes to mind is, moving the find call up decreases the time we keep the pair of unit registe