@jkeiser 2019-09-09
Oh I also think I didn't need to copy the trailing if statement 4 times, we should delete those before merge.