@jkeiser 2019-09-09
And finally, I haven't tested if we get similar benefit with different unrolling, ie 2 or 8 instead of 4