@jkeiser 2019-09-09
I also tried moving find whitespace and structural up above the mask processing, on the theory that we could discard the input registers even earlier, and the benefit disappeared. It seems like some sort of optimizer sweet spot.