@jkeiser 2019-09-10
I just updated this: I think the `if (idx < len)` blocks would lead to stack corruption depending on the length. The updated version fixes that. Note: Even the final iteration of the loop needed to be unrolled for the performance difference to stick (it was faster, but not that much faster).