@lemire 2019-09-10
Thanks. Let us figure out where the gains are coming from. I think that the gains are relatively large and not easily explained by the mere overhead of the loop. Something else might be happening and it would be instructive to figure it out. 

In an ideal world, we would have model of what is happening that we could explain (so that people know why the loop is unrolled).