@jkeiser 2019-09-11
Agreed.

New result: unrolling 2 iterations of loop gives just as much improvement as 3 or 4. As far as I can tell, going from 2->4 doesn't even improve it a little bit (if it does, it's within the noise limit). 8 starts to degrade it again, but I'm guessing we're hitting instruction cache limits again.

I'm super surprised that that benefit doesn't increase after unrolling > 2. I assumed it had something to do with the two iterations being optimized together.