@jkeiser 2019-09-11
After experimenting with more accurate numbers, I think the two changes are mostly orthogonal, and they have to be done together to show up against the full run, while they show clear benefit separately when you run just stage 1.

Hypotheses:

* **unrolling:** Gives about 3% improvement in stage 1 by itself.. I don't think we need to explain this one any further than inlining being good: we *know* we have data dependencies between one loop and another, even just in flatten_bits. 3 iterations is peak for the current code on my machine, but I'd leave it at 2 to hopefully fit better in the caches on smaller machines.
* **flatten_bits:** Gives about 1.5% improvement in stage 1 by itself. Less sure on this one. Moving flatten_bits down is likely related to instruction latency and data dependency again. Perhaps find_whitespace_and_structurals also spends time on some latent SIMD instructions, so flatten_bits can hide even more latency?

The experimental data:

* baseline: 1 iteration - 4.84 GB/s, 2 iterations - 4.99 GB/s, 4 iterations - 5.04 GB/s
* move flatten_bits: 1 iteration - 4.92 GB/s, 2 iterations - 5.13 GB/s, 4 iterations - 5.11 GB/s
* move flatten_bits even further down: 1 iteration - 4.13 GB/s, 2 iterations - 4.23 GB/s, 4 iterations - 4.19 GB/s

This is on my AVX-2 machine, and was heavily controlled:
* I ran stage 1 only to make the difference more pronounced.
* Variations for move of flatten bits below find_whitespace_and_structurals, as well as even further down to the bottom of find_structural_bits_64
* Variations of each with 1, 2, and 4 unrolled iterations per loop cycle
* Each performance run was checkperf against baseline (no modifications, 1 iteration)
* Threw out runs where the baseline didn't match 4.84, on the theory that the machine was busy during this run
* Picked the high number each time
* Ran multiple iterations of checkperf to ensure I got the same number (within 0.005 GB/s)