@jkeiser 2019-09-11
All right, I've gone as far as I can go with experiments, I think. I found another 3% improvement while I did that. Here is the series of steps to achieve a 9-10% gain across stage 1 (3.5-4% total improvement including stage 2) on AVX-2:

1. **3% gain:** Unroll the loop (2 iterations). Doesn't make a difference whether you copy/paste the functions or call one function twice in a row, but copy/paste the functions so you can do step 5 here. *Theory:* optimizer is just happier when the loop is fused together.
2. **3% gain:** Run `flatten_bits` after `find_whitespace_and_structurals`. *Theory:* find_whitespace_and_structurals has latency issues just like the clmul in `find_quote_mask_and_bits`, but flatten_bits gives it time to finish before the results are needed.
3. **1.5% gain:** Run the UTF-8 check *after* find_whitespace_and_structurals (and after flatten_bits). *Theory:* structural+clmul is bigger than anyone thought :)
4. **1.5% gain:** Run the UTF-8 check after find_whitespace_and_structurals but *before* flatten_bits. *Theory:* I'm stumped here. I have no idea why the ordering of the UTF-8 check and flatten_bits matters. Neither of them has any dependencies at all on the surrounding function calls. I suppose utf8_state shares an input with find_whitespace_and_structurals, so maybe it's able to access it faster? Maybe the fact that it's doing SIMD and flatten_bits isn't, causes the processor to be happier because it doesn't like switching between the instruction sets?