@lemire 2019-09-12
@jkeiser Welcome in the "obsessive-compulsive performance geek club".

It is a given that we will merge this.