@jkeiser 2019-09-13
OK! I got my Linux box up with this code and discovered some useful information. This does appear to be about branch prediction, but wildly enough, it's the *frequency* of branch misprediction that goes down. Specifically, the PR cuts the number of mispredicted branches by 40% (from 1060 -> 620). I don't have any explanation for this at all. I'll check what happens when you do the changes piecemeal, now that I'm armed with these metrics.

- **Master:**
  ```
  stage 1 instructions:    1790260 cycles:     487792 (99.83 %) ins/cycles: 3.67 mis. branches:        601 (cycles/mis.branch 810.30) cache accesses:      29879 (failure          4)
  stage 1 runs at 0.77 cycles per input byte.
  ```
- **PR:**
  ```
  stage 1 instructions:    1720956 cycles:     527807 (99.84 %) ins/cycles: 3.26 mis. branches:       1062 (cycles/mis.branch 496.83) cache accesses:      29986 (failure         27)
  stage 1 runs at 0.84 cycles per input byte.
  ```

BTW: when you only run stage 1, the numbers reported by `parse` for stage 1 are *way* different, way lower. I'm guessing there's an issue with the algorithm, but maybe there's genuinely some impact of stage 2 on the number of branch misses in stage 1--like it is hitting a limit on the number of branches the BPU can keep track of?

- **Master with both stage 1 and stage 2**
  ```
  stage 1 instructions:    1720956 cycles:     540511 (51.81 %) ins/cycles: 3.18 mis. branches:       1747 (cycles/mis.branch 309.31) cache accesses:      30040 (failure         22)
  stage 1 runs at 0.86 cycles per input byte.
  stage 2 instructions:    1597975 cycles:     501823 (48.11 %) ins/cycles: 3.18 mis. branches:       1278  (cycles/mis.branch 392.50)  cache accesses:      54107 (failure         94)
  stage 2 runs at 0.79 cycles per input byte and 9.08 cycles per structural character.
  all stages: 1.65 cycles per input byte.
```